+++ types/index.d.ts
@@ -11,26 +11,9 @@ interface IPaths {
   appNodeModulesBin: string;
   appNodeModulesCache: string;
   appDotEnv: string;
-  appTestSetupFiles: string[];
-}
-
-interface ISetupFilesPath {
-  path: string;
-  target: string;
 }
 
 interface ITestConfig {
   isWatch: boolean;
   isCoverage: boolean;
-}
-
-interface ILintestInfo {
-  package: string;
-  version: string;
-  rootPath: string;
-}
-
-interface IOwnPackageJson {
-  name: string;
-  version: string;
 }