+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable version changes would be documented in this file.
 
+## 1.2.0
+> Minor version updated at 2020.05.19
+- Chores
+  - Forked to `@lintest/cli` app for enterprise usage
+  - Update dependencies
+- New Features
+  - Adds `export debug` command to refer to the configuration file
+- Deprecated: Writing cache files in `node_modules` cache directory
+
 ## 1.1.0
 > Minor version updated at 2020.05.14
 - New Features: Adds `export` command for reference in application